fixed mockito import removed Position

+++ src/test/java/run/RobotTest.java
@@ -10,8 +10,8 @@ public class RobotTest {
     @Test
     public void SimpleRobotMovementTest() {
         Robot robot = new Robot(null);
-        float posX = robot.getPosX();
-        float posY = robot.getPosY();
+        int posX = robot.getPosX();
+        int posY = robot.getPosY();
         robot.setPos(posX+1,posY+1);
         assertEquals(robot.getPosX(), posX+1, 0);
         assertEquals(robot.getPosY(), posY+1, 0);
